remove configchecker
